libMcClient: Add 64-bit compatibility

+++ mobicore/daemon/ClientLib/ClientLib.cpp
@@ -392,7 +392,7 @@ __MC_CLIENT_LIB_API mcResult_t mcOpenSession(
         SEND_TO_DAEMON(devCon, MC_DRV_CMD_OPEN_SESSION,
                        session->deviceId,
                        *uuid,
-                       (uint32_t)(tci) & 0xFFF,
+                       (uintptr_t)(tci) & 0xFFF,
                        (uint32_t)handle,
                        len);
 
@@ -574,7 +574,7 @@ __MC_CLIENT_LIB_API mcResult_t mcOpenTrustlet(
                        session->deviceId,
                        spid,
                        (uint32_t)tlen,
-                       (uint32_t)(tci) & 0xFFF,
+                       (uintptr_t)(tci) & 0xFFF,
                        (uint32_t)handle,
                        len);
 
@@ -1219,7 +1219,7 @@ __MC_CLIENT_LIB_API mcResult_t mcMap(
                        session->sessionId,
                        (uint32_t)bulkBuf->handle,
                        (uint32_t)0,
-                       (uint32_t)(bulkBuf->virtAddr) & 0xFFF,
+                       (uintptr_t)(bulkBuf->virtAddr) & 0xFFF,
                        bulkBuf->len);
 
         // Read command response
@@ -1243,7 +1243,7 @@ __MC_CLIENT_LIB_API mcResult_t mcMap(
         RECV_FROM_DAEMON(devCon, &rspMapBulkMemPayload);
 
         // Set mapping info for internal structures
-        bulkBuf->sVirtualAddr = (void *)rspMapBulkMemPayload.secureVirtualAdr;
+        bulkBuf->sVirtualAddr = (void *)(uintptr_t)rspMapBulkMemPayload.secureVirtualAdr;
         // Set mapping info for Trustlet
         mapInfo->sVirtualAddr = bulkBuf->sVirtualAddr;
         mapInfo->sVirtualLen = bufLen;
@@ -1311,7 +1311,7 @@ __MC_CLIENT_LIB_API mcResult_t mcUnmap(
         SEND_TO_DAEMON(devCon, MC_DRV_CMD_UNMAP_BULK_BUF,
                        session->sessionId,
                        handle,
-                       (uint32_t)(mapInfo->sVirtualAddr),
+                       (uintptr_t)(mapInfo->sVirtualAddr),
                        mapInfo->sVirtualLen);
 
         RECV_FROM_DAEMON(devCon, &mcResult);
